better statistical dealy
git-svn-id: http://chara.hi.gemini.edu/svn/DRSoftware/gemini_python/trunk@3461 e04282df-fb77-4b70-bf7a-c1223586725c

+++ astrodata_Gemini/RECIPES_Gemini/primitives/primitives_GEMINI.py
@@ -91,7 +91,7 @@ class GEMINIPrimitives(BookkeepingPrimitives,DisplayPrimitives,
                         "datalabel": "GN-2012B-Q-0-000-%0.3d" % random.randint(1,999),
                         "filename": filename})
             delt = (now_lt - now.replace(hour=0,minute=0) )
-            nowsec = delt.days*86400 + delt.seconds 
+            nowsec = float(delt.days*86400 + delt.seconds)
                 
             return (mtd, nowsec)
         from time import sleep
